fix - Update tax rate methods to return BigDecimal instead of double for consistency

+++ src/main/java/co/invest72/investment/domain/Taxable.java
@@ -9,5 +9,5 @@ public interface Taxable {
 
 	String getTaxType();
 
-	double getTaxRate();
+	BigDecimal getTaxRate();
 }
