+++ src/main/java/co/invest72/investment/domain/tax/TaxBenefit.java
@@ -29,7 +29,7 @@ public class TaxBenefit implements Taxable {
 	}
 
 	@Override
-	public double getTaxRate() {
-		return taxRate.getValue().doubleValue();
+	public BigDecimal getTaxRate() {
+		return taxRate.getValue();
 	}
 }
